refactor: NewsApi
